chore: move shm_open_anon to common
move shm_open_anon to common
