+++ src/CMakeLists.txt
@@ -4,6 +4,7 @@ set(DIM_DATA_DIR ${CMAKE_INSTALL_DATAROOTDIR}/dim)
 set(DIM_ADDON_INFO_DIR ${DIM_DATA_DIR}/addon)
 set(DIM_ADDON_DIR ${CMAKE_INSTALL_LIBDIR}/dim)
 
+add_subdirectory(common)
 add_subdirectory(wl)
 add_subdirectory(imwl)
 add_subdirectory(dimwl)
