+++ src/common/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(shm_open_anon OBJECT
+    shm_open_anon.h
+    shm_open_anon.c
+)
